入力クラスの責務が単一になるように変更
