+++ 08.ls_object/lib/command_line_arguments.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+class CommandLineArguments
+  def initialize(argv = ARGV)
+    args = parse(argv)
+    @options = args[:options]
+    @path = args[:path]
+  end
+
+  attr_reader :options, :path
+
+  private
+
+  def parse(argv)
+    options = argv.getopts('arl')
+    path = argv[0] || Dir.getwd
+    { path:, options: }
+  end
+end
